split app.jsx into multiple components
